fix(ev): use command ls
This commit replaces raw "ls" with "command ls". "ls" may be overridden
by user functions with default (such as -lG). This breaks ev which
relies on a specific output format.

The same outcome could be achieved with find, but that complicates
things when dealing with symlinks and directories. Find would be more
portable, but more clunky.